Refactor PartitionHandler
